Implement serialization for CompatibleX types
To enable a graceful roll-out, it's important that the CompatibleX types
*serialize* to the v5 format, in case there is a service *consuming*
these APIs without being upgraded to the compatible types.

For now, this makes the v6 types purely a runtime layer. Every API
response, network message, or value stored in dynamodb should be
unimpacted.

We can make sure that every service is deployed with these types, and
start introducing these values into everything.

Eventually, once everything is using the CompatibleX types, we can switch
the serialization to use v6 types instead; since everything is upgraded,
it will be able to handle the new format seamlessly, and we won't have
to worry about the order we upgrade things.