refactor: drop v5 support
This PR completely removes the compatibility and v5 support. Nobody on
any public network is still using a v5 Ogmios by this time as it's no
longer compatible with those networks. As such, I've taken a hatchet to
the code and ripped it out. This also renames the entire package back to
simply `github.com/SundaeSwap-finance/ogmigo`, dropping the v6 postfix.

BREAKING CHANGES: v5 and compatibility support removed, package renamed

Signed-off-by: Chris Gianelloni <wolf31o2@blinklabs.io>