refactor: restore v5 support (#63)
This is an almost revert of 2ee5ca0 (#51) to restore the `v5` variant to allow for migration from the old types to the new types. This also includes the changes made to the `v6` branch ahead of `main` so this should bring them back in line and should allow users of the `v6` branch to use releases from `main` going forward.

Signed-off-by: Chris Gianelloni <wolf31o2@blinklabs.io>